+++ entrypoint.sh
@@ -1,0 +1,22 @@
+#!/bin/bash
+
+set -Eeuo pipefail
+
+function print_help {
+    echo "Available options:"
+    echo " start commands (Rasa Core cmdline arguments) - Start Rasa Core Action Server"
+    echo " help                                         - Print this help"
+    echo " run                                          - Run an arbitrary command inside the container"
+}
+
+case ${1} in
+    start)
+        exec python -m rasa_core_sdk.endpoint "${@:2}"
+        ;;
+    run)
+        exec "${@:2}"
+        ;;
+    *)
+        print_help
+        ;;
+esac